pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qocc"
version = "0.1.0"
description = "Quantum Observability + Contract-Based Compilation — trace-first, vendor-agnostic layer for quantum workflows"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [{ name = "QOCC Contributors" }]
keywords = ["quantum", "observability", "compilation", "contracts", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "click>=8.1",
    "jsonschema>=4.20",
    "numpy>=1.24",
    "pydantic>=2.5",
    "rich>=13.0",
]

[project.optional-dependencies]
qiskit = ["qiskit>=1.0"]
cirq = ["cirq-core>=1.3"]
qec = ["stim>=1.12", "pymatching>=2.0"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]
all = ["qocc[qiskit,cirq,dev]"]

[project.scripts]
qocc = "qocc.cli.main:cli"

[project.urls]
Repository = "https://github.com/zerocool26/Quantum-Observability-Contract-Compilation-QOCC-"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
